develop 2: correction of remarks from Riviera
