+++ README.md
@@ -1,1 +1,71 @@
-"# Diplom_2" 
+## **Projekt Diplom_2**
+
+### _Описание проекта_
+
+Этот проект представляет собой автоматизированное тестирование API Stellar Burgers.
+
+## **Технологии**
+Технология  |	Версия
+------------|-----------
+Java	    |   11
+Maven	|       3.9.0
+JUnit	|       4.13.2
+Rest Assured	|5.3.1
+Allure	        |2.15.0
+Lombok	        |1.18.24
+Gson	        |2.10.1
+
+## **Настройка**
+Для запуска проекта необходимо установить следующие зависимости:
+
+Технология  |	Версия
+------------|-----------
+Java	    |   11
+Maven	|       3.9.0
+JUnit	|       4.13.2
+Rest Assured	|5.3.1
+Allure	        |2.15.0
+Lombok	        |1.18.24
+Gson	        |2.10.1
+
+## **Запуск**
+
+Для запуска проекта используйте команду:
+
+`mvn clean test`
+
+Эта команда запустит тесты и сгенерирует отчет Allure.
+
+## **Проверка**
+
+Для проверки проекта используйте команду:
+
+`mvn allure:serve`
+
+Эта команда откроет сайт с отчетом
+
+## **Структура проекта**
+`src/main/java` - исходный код проекта.
+
+`src/test/java` - тесты проекта.
+
+`pom.xml` - файл конфигурации Maven.
+
+## **Зависимости**
+
+Зависимость |	Версия
+------------|-----------
+allure-junit4 |	2.15.0
+allure-rest-assured |	2.15.0
+rest-assured |	5.3.1
+selenium-java |	3.141.59
+juni |t	4.13.2
+webdrivermanager |	4.4.3
+lombok |	1.18.24
+gson |	2.10.1
+
+## **Плагины**
+Плагин |	Версия
+------------|-----------
+maven-surefire-plugin |	2.22.2
+allure-maven |	2.10.0
